+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 =====
 
 *   (improvement) Make `searchable` configurable.
+*   (internal) Mark functions as `@internal`.
+*   (internal) Remove obsolete function.
 
 
 2.1.1